requirements.txt
Adafruit-GPIO
Adafruit-PureIO
Nabaixin-TFOLED
Pillow
RPi.GPIO==0.7.1
spidev
psutil
